backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Greg Baranov <gbaranov99@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
djangorestframework = "^3.10"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.1"
django-rest-auth = "^0.9.5"
django-allauth = "^0.40.0"
gunicorn = "^19.9"
django-filter = "^2.2"
djangorestframework_simplejwt = "^4.3"
django-mptt = "^0.10.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
